requirements.txt
amqp==2.6.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
billiard==3.6.3.0
black==19.3b0
bleach==3.1.5
boto3==1.14.46
botocore==1.17.46
celery==4.4.7
celery-prometheus-exporter==1.7.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-cors-headers==3.4.0
django-extensions==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
docutils==0.15.2
drf-yasg==1.17.1
entrypoints==0.3
factory-boy==3.0.1
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
iniconfig==1.0.1
ipdb==0.13.3
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
pycurl==7.43.0.5
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-decouple==3.3
python-json-logger==0.1.11
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
